+++ packages/openapi3-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # API Elements: OpenAPI 3 Parser Changelog
 
+## TBD
+
+### Enhancements
+
+- Support for 'Schema Object' `additionalProperties` when the value is boolean.
+
 ## 0.16.0 (2021-02-23)
 
 ### Enhancements